+++ tests/test_core_exclude.py
@@ -1,4 +1,3 @@
-import pytest
 from urlchecker.core.exclude import excluded